Rust migration Phase 8 M11-M12: xtask shader pipeline and phase exit (#40)
* Rust migration Phase 8 M11: xtask shader pipeline and embedded pak

Task plan D8 (docs/ai/plans/rust-conversion-phase-8.md), ADR-015, ADR-003.

Make rust/xtask a lib+bin crate that reproduces meson's shader pipeline
(glslangValidator/spirv-opt with the same flags, 61 jobs), the bintoc text
format with its deflate step (miniz_oxide, TDEFL_MAX_PROBES_MASK) and the
mkpak PACK writer. Under the `render` feature quake-capi's new build.rs
compiles the shaders and builds the embedded pak into OUT_DIR and bakes them
in with include_bytes!; gl_rmisc.rs reads the embedded table instead of the
122 C externs and embedded_pak.rs exports the vkquake_pak symbols for
common_fs.c. meson.build skips the shader, mkpak and bintoc -c custom targets
under use_rust_render and passes the tool paths to cargo; the C-only build
and the native bintoc/mkpak tools are unchanged (NG3, deletion in Phase 9/10).

Verification: scripts/harness/xtask_diff.py compares every .spv/.c, the pak
and embedded_pak.c against build-c (124 byte-identical, AC10, release
buildtype); quake-ctest's xtask_differential pins deflate_raw against the
in-tree tdefl (real pak identical; the miniz_oxide <=32-byte stored-block
divergence is pinned and recorded); xtask's job_table_covers_every_shader
ties the job table to quake-render's Shader::ALL. The rust.yml lint jobs
install the SDK tools and run the xtask tests; the Linux harness job runs
xtask_diff.py. Plan amendment log, AC10, evidence and gate updated.


* Rust migration Phase 8 M12: phase exit

Run the full ADR-019 suite on the Windows dev box and record the Phase 8
exit block in the ROADMAP (status, evidence, exit-criterion deviations,
the first ADR-004 unsafe inventory, deletions deferred to Phase 9 with
the switches). Two findings from the debug-buildtype render leg:

- r_brush.rs: build_surface_display_list computed the C
  `sizeof(glpoly_t) + (lnumverts - 4) * VERTEXSIZE * sizeof(float)` term
  with a usize add of a wrapped negative product, which is bit-identical
  in release but panics with `attempt to add with overflow` in debug for
  3-edge surfaces (M8 regression; the debug leg was last run at M6/M7).
  Use signed arithmetic and cast once.
- render_corpus.py --validation now fails unless the run printed
  `Using VK_LAYER_KHRONOS_validation`: the layer is only requested under
  _DEBUG (gl_vidsdl.c, engine-debug in quake-render), so the release-leg
  "validation clean" claims of M2-M11 were vacuous. The phase-exit
  validation evidence is build-c-debug vs build-rs-debug.


* Record the PR #40 CI evidence for Rust migration Phase 8 M11-M12

Replace the "CI legs pending the push" wording in the Phase 8 task plan
and ROADMAP exit block with the PR #40 results: every running check
green on the first run (Linux xtask_diff and lavapipe stability legs,
TSan, miri, loom, fuzz smoke, the three-OS harnesses and builds); cargo
audit and the f32 sweep remain schedule-only. Docs only.


* Fix the PR #40 review findings and the sv_phys fixture flake (Phase 8 M12)

Review: xtask_diff.py resolved --out relative to the cargo working
directory, so a relative --out landed under rust/ and --out <build-dir>
would have overwritten the oracle; make both paths absolute, reject
--out == --build-dir, and pass --locked to cargo run. Correct the plan's
Tools::find precedence (env vars, PATH, then $VULKAN_SDK/bin). The
LD_LIBRARY_PATH suggestion is declined: glslangValidator and spirv-opt
do not load the Vulkan loader.

CI: the ubuntu unit+differential job aborted in sv_phys_differential
with "NUM_FOR_EDICT: bad pointer". run_raise_case raises inside
SV_Physics, whose Host_Error longjmp skips the epilogue that clears the
push-grid/pushable-cache statics and the alloc hook, so the next test's
sv_fastpushmove 0 tick walked pointers into a freed arena. Heal the
statics in the fixture with one complete fast-pushers tick per side and
clear the hook slot; the engine is unchanged (C stays the oracle).
Record the dispositions, root cause, confirmation and a residual note in
the task plan amendment log and ROADMAP.


* Address the PR #40 owner review on the xtask gate and sv_phys fixture (Phase 8 M12)

Seven findings from the second review, all accepted:

- xtask_diff.py compares the .spv name sets in both directions, refuses a
  debug buildtype up front via meson-info, documents that both sides must
  run the same glslang/spirv-opt binaries, removes its default temp --out
  on success and names it on failure, and points the embedded_pak.c
  failure at the pinned miniz_oxide/tdefl divergence when the pak itself
  is identical.
- sv_phys_differential.rs resets the pusher statics from a Drop guard
  created before the raise loop, so a real raise-case mismatch reports as
  one failed test instead of an unarmed NUM_FOR_EDICT abort in the next
  test (verified with a deliberately broken needle: 27 passed, 1 failed).
- xtask compile_all bounds its thread pool by Cargo's NUM_JOBS under
  build.rs; deflate_raw doubles the output buffer on every Okay, as
  tdefl_compress_mem_to_heap does.
- readme.md records that cargo check/clippy -p quake-capi --features
  render needs glslangValidator and spirv-opt.
- rust.yml's flattened tar --wildcards line restored to the multi-line
  form of build-linux.yml.

The plan's amendment log has the row.


---------

+++ rust/quake-capi/src/r_brush.rs
@@ -1841,9 +1841,11 @@ unsafe fn build_surface_display_list(fa: *mut MSurface) {
     unsafe {
         let pedges: *mut MEdge = (*CURRENTMODEL).edges;
         let lnumverts = (*fa).numedges;
+        // A 3-edge surface makes the C `(lnumverts - 4) * ...` term negative;
+        // signed arithmetic keeps the debug-profile overflow check quiet.
         let poly = c::Mem_Alloc(
-            core::mem::size_of::<GlPoly>()
-                + ((lnumverts as isize - 4)
+            (core::mem::size_of::<GlPoly>() as isize
+                + (lnumverts as isize - 4)
                     * VERTEXSIZE as isize
                     * core::mem::size_of::<f32>() as isize) as usize,
         )
